requirements.txt
pyyaml
python-dotenv
google-genai
mistralai
playwright
pandas
numpy
matplotlib
seaborn
